Clean up Cluster Creation
Separates the massive blob to render the modal into its own file to
promote cleanliness.
